[reorg] Move GSUB.hh up one level and change GSUB namespace to GSUB_impl.
