+++ src/Makefile.sources
@@ -118,7 +118,7 @@ HB_BASE_sources = \
 	OT/Layout/GSUB/ExtensionSubst.hh \
 	OT/Layout/GSUB/SubstLookupSubTable.hh \
 	OT/Layout/GSUB/SubstLookup.hh \
-	OT/Layout/GSUB/GSUB.hh \
+	OT/Layout/GSUB.hh \
 	OT/Layout/GPOS.hh \
 	OT/Layout/GPOS/CursivePosFormat1.hh \
 	OT/Layout/GPOS/MarkLigPos.hh \
