+++ src/OT/Layout/GSUB/AlternateSubstFormat1.hh
@@ -6,7 +6,7 @@
 
 namespace OT {
 namespace Layout {
-namespace GSUB {
+namespace GSUB_impl {
 
 struct AlternateSubstFormat1
 {
