+++ src/OT/Layout/GSUB/MultipleSubstFormat1.hh
@@ -6,7 +6,7 @@
 
 namespace OT {
 namespace Layout {
-namespace GSUB {
+namespace GSUB_impl {
 
 struct MultipleSubstFormat1
 {
