+++ src/OT/Layout/GSUB/SingleSubstFormat2.hh
@@ -5,7 +5,7 @@
 
 namespace OT {
 namespace Layout {
-namespace GSUB {
+namespace GSUB_impl {
 
 struct SingleSubstFormat2
 {
