+++ src/meson.build
@@ -122,7 +122,7 @@ hb_base_sources = files(
   'OT/Layout/GSUB/ExtensionSubst.hh',
   'OT/Layout/GSUB/SubstLookupSubTable.hh',
   'OT/Layout/GSUB/SubstLookup.hh',
-  'OT/Layout/GSUB/GSUB.hh',
+  'OT/Layout/GSUB.hh',
   'OT/Layout/GPOS.hh',
   'OT/Layout/GPOS/CursivePosFormat1.hh',
   'OT/Layout/GPOS/MarkLigPos.hh',
